fix: `ExecuteSwitch` method now powers display on
PowerOn/Off methods have been refactored to not rely on feedbacks and to
make it more explicit when they will be skipped. Also modified them to
not send the power on/off commands if the method is called while the
display is changing states.

There were 2 factories in the module that have now been merged. Also
removed extraneous constructors that really weren't necessary.

Warmup/Cooldown times are now configurable in milliseconds via the device properties. If
not set, the times default to 5 seconds.